# Conflicts:
#	src/main/java/org/webeng/collector_site/controller/collezioni/ShowCollezione.java

+++ src/main/java/org/webeng/collector_site/controller/collezioni/DeleteDiscoCollezione.java
@@ -51,9 +51,15 @@ public class DeleteDiscoCollezione extends CollectorsBaseController {
         Disco disco=((CollectorsDataLayer) request.getAttribute("datalayer")).getDiscoDAO().getDisco(Integer.parseInt(request.getParameter("id_disco")));
         request.setAttribute("collezione",collezione);
         request.setAttribute("disco",disco);
-        ((CollectorsDataLayer) request.getAttribute("datalayer")).getDiscoDAO().deleteDisco(collezione,disco);
+        if(Utility.getUtente(request,response).equals(collezione.getUtente())){
+            ((CollectorsDataLayer) request.getAttribute("datalayer")).getDiscoDAO().deleteDisco(collezione,disco);
+            response.sendRedirect(request.getHeader("Referer"));
+        }
+        else {
+            response.sendRedirect("/home");
+        }
 
-        response.sendRedirect(request.getHeader("Referer"));
+
     }
 
     private void action_anonymous(HttpServletRequest request, HttpServletResponse response) throws IOException  {
